+++ Packages/webxr-interactions/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - SceneHitTest now updates the pose of the originTransform instead of WebXRManager transform.
+- Mapping of WebXR Tracked Display device.
 
 ## [0.20.0] - 2023-12-18
 ### Added